1) Removed race_setup and race_mode data structures. All this    information is now only managed by the race_manager, no    more in-between objects to transfer information along. 2) The scores for grand prix are now defined in the stk_config.dat    file (10, 8, 6, 5, 4, .., 1, 0, 0) points 3) Bugfix: unlock information wasn't saved anymore. Added specific    saving after unlocking, plus re-inserted the 'generic' save    at the end of STK again. 4) bugfix/work around: Visual Studio complains about incompatible    iterators in sdldrv - apparently caused by using erase, and    then keep on using the iterator. 5) Fixed bug when running a race in a GP again (scores/times    were added each time).
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk/supertuxkart@1681 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/user_config.cpp
@@ -732,7 +732,7 @@ UserConfig::writeStickConfigs(lisp::Writer *writer)
 	
 	writer->beginList("stick-configs");
 	
-	count = m_stickconfigs.size();
+	count = (int)m_stickconfigs.size();
 	writer->write("count", count);
 	
 	for (int i = 0; i < count; i++)
